Day4&Day5/requirements.txt
Django==4.2.7
openai==1.3.7
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.5
chromadb==0.4.18
sentence-transformers==2.2.2
numpy>=1.26.0
pandas>=2.1.0 